Allow get_per_commitment_point to fail.
This changes `ChannelSigner::get_per_commitment_point` to return a
`Result<PublicKey, ()`, which means that it can fail. Similarly, it changes
`ChannelSigner::release_commitment_secret`. In order to accomodate failure at
the callsites that otherwise assumed it was infallible, we cache the next
per-commitment point each time the state advances in the `ChannelContext` and
change the callsites to instead use the cached value.