For tests, use a testing-specific class so we can tweak its behavior.
This wraps the `EnforcingSigner` implementation with a `TestChannelSigner` and
fixes the usage accordingly. While `TestChannelSigner` currently forwards all
the handlers unchanged to the `EnforcingSigner`, the idea would be to provide
some additional configuration that would allow the `TestChannelSigner` to
exhibit different behavior -- for example, returning errors for fallible
methods.

Remove accidental test.