Check if HTLC is worth claiming onchain when failing from monitor event
In previous commits, we added a config flag to allow a user to control a
multiplier on a threshold for whether to fail back an HTLC early if
we're waiting on a downstream confirmation, depending on how much we'd
save if we ended up successfully claimed on-chain. This commit adds the
logic to calculate this threshold, and adds the check when failing HTLCs
back from monitor events (although this won't do anything until we
actually generate new monitor events for this situation in upcoming
commits).