Handle signer unavailable during inbound channel creation
When accepting a new inbound channel, we need to acquire the first
per-commitment point. If the signer is not immediately available to do so,
then unwind and allow for retry later.

This changes the next_per_commitment_point to be an Option<PublicKey> which
will only be None while waiting for the first per-commitment point.