Allow release_commitment_secret to be fallible
Rather than assume that we can release the commitment secret arbitrarily, this
releases the secret eagerly: as soon as the counterparty has commited to the
new state. The secret is then stored as part of the channel context and can be
accessed without a subsequent API request as needed.

Also adds support for serializing and deserializing the previous commitment
secret and next per-commitment point with the channel state.