Keep Tokio running during read_Event
This change keeps Tokio from stalling while handling a long-running
`read_event`. It uses `tokio::task::spawn_blocking` to use Tokio's blocking
thread pool to process the task, which allows the main event loop to continue
to run even if `read_event` results in I/O, etc.

On OS/X, the kevent implementation seems to have an issue where
`spawn_blocking` is insufficient. It's also necessary to continue to `yield`
on the current task. So... conditionally do this for OS/X.