Correct syn pinning on cargo 1.48
Sadly the pinning introduced in 050f5a90297678f2cad36feee9a1db2367e
was brittle in the face of any further syn updates, and has already
broken.

Here we fix it by looking up the actual version of syn to pin.

Note that this dependency is somewhat nonsense as its actually only
a `criterion` dependency, pulled in even though we haven't set the
bench flag (as we aren't yet using `resolver = 2`).